backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
python-dotenv
requests
supabase
pytest
